default.latex: wrong beamer color in (sub)section page
The default.latex template defines "part title", "section title" and "subsection title" via \setbeamertemplate. But for "section title" and "subsection title", the beamercolorbox also used the "part title" parameter instead of the "(sub)section title". This had the effect that for (sub)section pages the color attributes of the part title were used.